requirements.simple.txt
# requirements.simple.txt - Version ultra-légère avec Google AI

# Dépendances existantes (gardez vos dépendances actuelles)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Base de données vectorielle légère
scikit-learn==1.3.2

# Traitement de texte
markdown==3.5.1
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
tiktoken==0.5.1

# Support Google AI (embeddings + génération)
google-generativeai==0.8.3

# Utilitaires
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0

# Monitoring (optionnel)
prometheus-client==0.19.0
        